requirements.txt
fasttext-wheel==0.9.2
openai==1.78.1
fastapi==0.110.2
uvicorn==0.34.2
python-dotenv==1.1.0
pandas==2.2.3
scikit-learn==1.6.1